generate-mf: Add option to use sha256
We still default to sha1, because there's no strong reason to switch to
sha256 at the moment, and if we do, we might break old versions of
VirtualBox of VMware that would not support it (although I have no idea
which versions would break).

For the background:
- the version 1.x of the OVF spec only mentions SHA1 as being supported
  in the manifest (cf. [1], check chapter 5 « OVF Packages »)
- the version 2.x of the OVF spec mentions both SHA1 and SHA256 (cf.
  [2])

But who cares about the spec anyway, at least not VMware Workstation 16:
when exporting to OVA, it gives you OVF v1 with SHA256 checksums.
VirtualBox 7 is a bit more behaved, it allows exporting to OVF v1, which
gives SHA1 checksums, and to OVF v2, which gives SHA256 checksums.

In any case, it seems that importing a OVF v1 with SHA256 checksums
works fine, for both VirtualBox and VMware (tested with versions given
above). So switching to SHA256 sums would not break recent versions of
those, but who knows, it might break older versions that didn't expect
SHA256 in the .mf file.

[1] https://www.dmtf.org/sites/default/files/standards/documents/DSP0243_1.1.1.pdf
[2] https://www.dmtf.org/sites/default/files/standards/documents/DSP0243_2.0.0.pdf