Skip root warning if inside container
The purpose of the warning is to tell users 'no need to be root, don't
use sudo'. However in container, it's often more convenient to be root
(ie. rootless container, running as root inside), in such case the
warning doesn't help, it's even probably misleading.

So let's try a bit harder.