Make sure to install kali-linux-firmware
So far, it was installed anyway, via the dependency chain:

  kali-linux-default -> kali-linux-headless -> kali-linux-firmware

However, what about user who don't install the default toolset
kali-linux-default? They would not have kali-linux-firmware installed,
and that's not intended. Having Wi-Fi firmware in the VM is useful, for
those who connect a USB Wi-Fi dongle and use it from within the VM.

So make sure to always install kali-linux-firmware.