Revert "No need to install usr-is-merged manually anymore"
This reverts commit 36a164a021aa885f962029dc1c34e413f2fc0970.

I was wrong, we still need to install usr-is-merged manually. The
debootstrap logs start with:

    excluding usr-is-merged as package is not in suite

So usr-is-merged is NOT installed by debootstrap. This is not due to
debootstrap itself though. This behaviour is specific to debos, cf.
https://github.com/go-debos/debos/commit/18998ffa.

Debos lists known recent Debian suites (sid, testing, bookworm...), and
assume the rest is either old suites, or derivatives based on old
suites, and therefore pass '--exclude=usr-is-merged' to debootstrap.

They basically take the opposite approach of debootstrap (which is to
list old suites, and assume the rests is recent suites). I suppose their
main use-case are derivatives based on older Debian suites.