Remove the CD/DVD from the VMware VM (thx X0RW3LL)
With kernel 6.5 (up to 6.5.5 it seems, didn't test anything after that),
VMware floods the logs with error messages related to VIDE (Virtual
IDE?), such as:

    In(05) vcpu-3 VIDE: (0x170) Unexpected IN for cmd 0xa3 len 2

It causes the VM to be laggy to the point of being completely unusable.

Thanks to investigations from X0RW3LL, it turns out that we can
workaround the issue by removing the CD/DVD device (which seems to be
the only IDE device).

So let's remove it for now until the issue is fixed.